+++ CHANGELOG.rst
@@ -7,7 +7,7 @@ Changelog
 * :issue:`16`: Add new method :meth:`pyvcf.VcfFrame.cfilter_empty`.
 * Add new methods :meth:`pyvep.filter_af/lof`.
 * Add ``matplotlib-venn`` package as dependency for plotting Venn diagrams.
-* Add new method :meth:`pyvcf.plot_comparison`.
+* Add new methods :meth:`pyvcf.plot_comparison/regplot`.
 * :issue:`17`: Add new method :meth:`pyvep.filter_biotype`.
 
 0.10.0 (2021-06-03)